+++ src/menu/joystick_options.c
@@ -1,6 +1,6 @@
 #include "joystick_options.h"
 
-#include "../controls/controller.h"
+#include "../system/controller.h"
 #include "../font/dejavusans.h"
 #include "../audio/soundplayer.h"
 #include "../savefile/savefile.h"
