+++ src/menu/menu_builder.c
@@ -1,7 +1,7 @@
 #include "menu_builder.h"
 
 #include "../util/memory.h"
-#include "../controls/controller.h"
+#include "../system/controller.h"
 #include "./translations.h"
 #include "../audio/soundplayer.h"
 #include "../util/time.h"
