+++ src/menu/save_game_menu.c
@@ -2,7 +2,7 @@
 
 #include "../savefile/savefile.h"
 #include "../levels/levels.h"
-#include "../controls/controller.h"
+#include "../system/controller.h"
 #include "../util/memory.h"
 #include "../audio/soundplayer.h"
 #include "./translations.h"
