+++ src/menu/savefile_list.c
@@ -1,7 +1,7 @@
 #include "./savefile_list.h"
 
 #include "../font/dejavusans.h"
-#include "../controls/controller.h"
+#include "../system/controller.h"
 #include "../util/rom.h"
 #include "../graphics/image.h"
 #include "../audio/soundplayer.h"
